requirements.txt
gherkin-official
nltk
openapi_spec_validator
textblob
inflect
pyyaml
python-dateutil
graphviz
